backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
qdrant-client==1.12.1
sentence-transformers==3.3.1
transformers==4.47.1
torch==2.5.1
pyannote.audio==3.3.2
faster-whisper==1.1.0
numpy==1.26.4
scipy==1.14.1
python-multipart==0.0.20
apscheduler==3.10.4
httpx==0.28.1
soundfile==0.12.1
librosa==0.10.2.post1
